Fixed bug by chaning numeric type of laser::Point to double. No idea what that might  break

+++ src/laser_utilities.h
@@ -74,17 +74,17 @@ namespace laser {
 	{
 	public:
 		Point() {}
-		Point(int x, int y)
+		Point(double x, double y)
 		{
 			m_x = x;
 			m_y = y;
 		}
 
-		int x() const {return m_x;}
-		int y() const {return m_y;}
+		double x() const {return m_x;}
+		double y() const {return m_y;}
 
-		void setX(int x) {m_x = x;}
-		void setY(int y) {m_y = y;}
+		void setX(double x) {m_x = x;}
+		void setY(double y) {m_y = y;}
 
 		Point scaled(double scale) const
 		{ return scaled(scale, scale); }
@@ -132,8 +132,8 @@ namespace laser {
 		}
 
 	protected:
-		int m_x;
-		int m_y;
+		double m_x;
+		double m_y;
 	};
 
 }
